Move helper functions from main to utils. Work on making stats global instead of per exam
