+++ views/quiz.py
@@ -1,5 +1,5 @@
 from flask import abort, Blueprint, redirect, render_template, request, session, url_for
-from sqlalchemy import func
+from utils import save_answer, session_data, random_id, sort_exam
 import models
 import random
 import re
@@ -36,9 +36,12 @@ def reset_stats_course(course):
 @quiz.route('/reset/<string:course>/<string:exam>/')
 def reset_stats_exam(course, exam):
     """Reset stats for a course"""
-    if 'exams' in session:
-        if exam and course + '_' + exam in session['exams']:
-            del session['exams'][course + '_' + exam]
+    course = models.Course.query.filter_by(code=course).first_or_404()
+    exam = models.Exam.query.filter_by(course=course, name=exam).first_or_404()
+    c_session = session_data('courses', course)
+    questions = {question.id for question in exam.questions}
+    c_session['answered'] = set(c_session['answered']) - questions
+    c_session['total'] = len(c_session['answered'])
     return redirect(url_for('quiz.exam', course=course, exam=exam))
 
 
@@ -67,15 +70,11 @@ def show_question(course, exam, id):
     exam = None
     if exam_name != 'all':
         exam = models.Exam.query.filter_by(course=course, name=exam_name).first_or_404()
-        # Shortened variable
-        c_session = session_data(session, 'exams', exam)
         # Only question from a specific exam
         num_questions = models.Question.query.filter_by(exam=exam).count()
         question = models.Question.query.filter_by(exam=exam).offset(id - 1).limit(1).first_or_404()
         reset_url = url_for('quiz.reset_stats_exam', course=course.code, exam=exam.name)
     else:
-        # Shortened variable
-        c_session = session_data(session, 'courses', course)
         # All questions
         num_questions = models.Question.query.filter_by(course=course).count()
         question = models.Question.query.filter_by(course=course).offset(id - 1).limit(1).first_or_404()
@@ -103,16 +102,9 @@ def show_question(course, exam, id):
                 if str(alternative.id) == answer:
                     context['success'] = alternative.correct
                     break
+            answered = save_answer(course, question.id, context['success'])
             # Checking if question has already been answered
-            if id not in c_session['answered']:
-                c_session['points'] += int(context['success'])
-                c_session['total'] += 1
-                c_session['answered'].append(id)
-                if context['success']:
-                    c_session['combo'] += 1
-                else:
-                    c_session['combo'] = 0
-            elif context['success']:
+            if answered and context['success']:
                 context['alerts'].append({
                     'msg': 'Du har allerede svart på dette spørsmålet så du får ikke noe poeng. :-)',
                     'level': 'info'
@@ -129,42 +121,5 @@ def show_question(course, exam, id):
     else:
         # Random order on questions
         random.shuffle(question.alternatives)
-    context['score'] = c_session
+    context['score'] = session_data('courses', course)
     return render_template('quiz/question.html', **context)
-
-
-def random_id(id=None, course=None, exam=None):
-    """
-        Returns a random id from questions that have not been answered.
-        Returns a random number if none available
-    """
-    if exam:
-        num_questions = models.Question.query.filter_by(exam=exam).count()
-        answered = session.get('exams', {}).get(exam.string, {}).get('answered', [])
-    elif course:
-        num_questions = models.Question.query.filter_by(course=course).count()
-        answered = session.get('courses', {}).get(course.string, {}).get('answered', [])
-    questions = set(range(1, num_questions + 1)) - set(answered) - {id}
-    if questions:
-        return random.choice(list(questions))
-    else:
-        # All questions have been answered
-        return random.randint(1, num_questions + 1)
-
-
-def session_data(session, index, model):
-    if index not in session:
-        session[index] = {}
-    if model.string not in session[index].keys():
-        session[index][model.string] = {'points': 0, 'total': 0, 'combo': 0, 'answered': []}
-    return session[index][model.string]
-
-
-def sort_exam(exam):
-    """Silly way to sort exams by V09, H09, V10, H10 etc."""
-    key = str(exam)
-    if re.match(r'^V\d{2}$', key):
-        key = key[1:3] + '1'
-    elif re.match(r'^H\d{2}$', key):
-        key = key[1:3] + '2'
-    return key
